Refetch program on run (path mode only)
Everytime the user hits the run button (whilst in path mode), refetch
the bundle from juttle-service and recieve any updates to the program.

As such add a bundle_id attribute (in path mode this is the path) to
tell if we are running a new bundle or updated bundle. If new bundle,
clear inputs; if updated bundle leave inputs be.

Support bundle_id from rendezvous service.

In addition: Move error rendering to its own view === less compilicated
run.js