Fix code toggle btn clearing view
If you already had a bundle going on, the code toggle btn was causing a
route action which was making the app clear the view. I've changed
things so we only navigate if runMode is not set. Otherwise the button
behaves like a regular btn, hiding and showing the code view.